Configure the kubelet to register CP nodes with the node-role.kubernetes.io/control-plane taint (#508)
This is necessary during disaster recovery scenarios, where we restore an etcd snapshot on a running cluster and we need to restart the kubelet to register an existing CP node to the recovered cluster state. Without this the CP node would get registered without the taint, allowing any Pod to get scheduled on it.

For context, normally this taint is added automatically by `kubeadm init` or `kubeadm join`, which will just proceed normally if it's already added by the kubelet before. Kubeadm is not involved during a DR scenario, since it's not designed to be re-run on an already bootstrapped node.

Related to https://github.com/giantswarm/giantswarm/issues/30439