Configure cilium to use the k8s hostname defined in cluster-info CM (#553)
Right now, cilium uses the external public hostname of the cluster to access the API server. This makes a critical internal component dependant on the baseDomain DNS infrastructure, which is sometimes managed by the customer, sometimes managed by us, and could be unreliable. Additionally it makes baseDomain more rigid and difficult to change on an existing cluster.

When setting `k8sServiceHost: auto`, the cilium helm chart will fetch the k8s API hostname from the `cluster-info` ConfigMap in the `kube-public` Namespace, which is set to an internal hostname or IP inside the provider network (e.g. the ELB hostname in AWS and Azure).